Factor out multiplication code
